long 타입으로 초단위 저장. 시간에 따라 3가지 형식으로 출력 (N s, N min, Nh N min)

+++ app/src/main/java/kr/cnu/ai/lth/adventuredesign/HistoryDBHelper.java
@@ -35,7 +35,7 @@ public class HistoryDBHelper extends SQLiteOpenHelper {
         onCreate(db);
     }
 
-    public long insertHistory(int detectCnt, int duration) {
+    public long insertHistory(int detectCnt, long duration) {
         SQLiteDatabase db = getWritableDatabase();
         ContentValues values = new ContentValues();
         values.put("DETECT_COUNT", detectCnt);
